+++ app/src/main/java/chiu/chingting/android_mvp/BaseFragment.java
@@ -1,4 +1,4 @@
-package chiu.chingting.calculator;
+package chiu.chingting.android_mvp;
 
 import android.app.Fragment;
 import android.os.Bundle;
