+++ app/src/main/java/chiu/chingting/android_mvp/BaseView.java
@@ -1,4 +1,4 @@
-package chiu.chingting.calculator;
+package chiu.chingting.android_mvp;
 
 public interface BaseView<T> {
     void setPresenter(T presenter);
