+++ app/src/main/java/chiu/chingting/android_mvp/GlobalConstance.java
@@ -1,4 +1,4 @@
-package chiu.chingting.calculator;
+package chiu.chingting.android_mvp;
 
 /**
  * Created by ting.chiu on 2016/12/12.
