+++ app/src/main/java/chiu/chingting/android_mvp/model/MemberInfo.java
@@ -1,4 +1,4 @@
-package chiu.chingting.calculator.model;
+package chiu.chingting.android_mvp.model;
 
 /**
  * Created by ting.chiu on 2016/12/9.
